fix: delete context trong create product
