+++ BabyStore/Pages/Admin/ProductManagement/Edit.cshtml
@@ -148,7 +148,7 @@
     <div class="container-fluid py-4">
         <div class="row">
             <div class="col-md-8">
-                <form method="post">
+                <form method="post" enctype="multipart/form-data">
                     <div class="card">
                         <div class="card-header pb-0">
                             <div class="d-flex align-items-center">
