+++ Services/Services.csproj
@@ -7,15 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-
-    <PackageReference Include="Microsoft.AspNetCore.Hosting" Version="2.2.7" />
-
-    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
-
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\Repos\Repos.csproj" />
   </ItemGroup>
